+++ packages/tinacms/src/toolkit/form-builder/form-builder.tsx
@@ -22,11 +22,7 @@ import { ResetForm } from './reset-form';
 import { CreateBranchModal } from './create-branch-modal';
 
 export interface FormBuilderProps {
-  form: {
-    tinaForm: Form;
-    activeFieldName?: string;
-    hoveringFieldName?: string;
-  };
+  form: { tinaForm: Form; activeFieldName?: string };
   hideFooter?: boolean;
   label?: string;
   onPristineChange?: (_pristine: boolean) => unknown;
@@ -213,7 +209,6 @@ export const FormBuilder: FC<FormBuilderProps> = ({
                     <FieldsBuilder
                       form={tinaForm}
                       activeFieldName={form.activeFieldName}
-                      hoveringFieldName={form.hoveringFieldName}
                       fields={fieldGroup.fields}
                     />
                   ) : (