+++ packages/@tinacms/mdx/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @tinacms/mdx
 
+## 2.0.3
+
+### Patch Changes
+
+- Updated dependencies [[`6c2c48a`](https://github.com/tinacms/tinacms/commit/6c2c48a9d869cb98e78fc656b986ecc244a5dafd)]:
+  - @tinacms/schema-tools@2.3.0
+
 ## 2.0.2
 
 ### Patch Changes