+++ packages/next-tinacms-azure/CHANGELOG.md
@@ -1,5 +1,19 @@
 # next-tinacms-azure
 
+## 8.0.1
+
+### Patch Changes
+
+- Updated dependencies [[`3fa216c`](https://github.com/tinacms/tinacms/commit/3fa216c110af6417e6b835c49a275b08981e028b), [`7352c96`](https://github.com/tinacms/tinacms/commit/7352c9660869d413a0f48d3fbb003d4c4c0e3d85)]:
+  - tinacms@3.3.1
+
+## 8.0.0
+
+### Patch Changes
+
+- Updated dependencies [[`6c2c48a`](https://github.com/tinacms/tinacms/commit/6c2c48a9d869cb98e78fc656b986ecc244a5dafd)]:
+  - tinacms@3.3.0
+
 ## 7.0.0
 
 ### Patch Changes